# backend/bin/Debug/net8.0/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Art Gallery",
      "description": "Backend service that provides resources for the beautiful art gallery.",
      "contact": {
        "name": "Tanay and Jeetasha",
        "email": "tanay4847.be22@chitkara.edu.in"
      }
    },
    "paths": {
...
      "/api/exhibitions": {
        "get": {
          "tags": [
            "Exhibition"
          ],
          "summary": "Retrieves all exhibitions.",
          "description": "Sample request:\r\n            \r\nGET /api/exhibitions",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Exhibition"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Exhibition"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Exhibition"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Exhibition"
          ],
          "summary": "Adds a new exhibition.",
          "description": "Sample request:\r\n            \r\nPOST /api/exhibitions\r\n{\r\n    \"name\": \"Impressionist Masterpieces\",\r\n    \"description\": \"A collection of Impressionist artworks from various artists.\",\r\n    \"place\": \"Art Gallery\",\r\n    \"date\": \"2024-05-01T00:00:00\"\r\n}",
          "requestBody": {
            "description": "The exhibition to add",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exhibition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exhibition"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Exhibition"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Created",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Exhibition"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Exhibition"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Exhibition"
                  }
                }
              }
            }
          }
        }
      },
      "/api/exhibitions/{id}": {
        "get": {
          "tags": [
            "Exhibition"
          ],
          "summary": "Retrieves a specific exhibition by its ID.",
          "description": "Sample request:\r\n            \r\nGET /api/exhibitions/1",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the exhibition to retrieve",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Exhibition"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Exhibition"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Exhibition"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Exhibition"
          ],
          "summary": "Updates an existing exhibition.",
          "description": "Sample request:\r\n            \r\nPUT /api/exhibitions/1\r\n{\r\n    \"id\": 1,\r\n    \"name\": \"Updated Impressionist Masterpieces\",\r\n    \"description\": \"An updated description of the Impressionist Masterpieces exhibition.\",\r\n    \"place\": \"Art Gallery\",\r\n    \"date\": \"2024-05-01T00:00:00\"\r\n}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the exhibition to update",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "description": "Updated information for the exhibition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exhibition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exhibition"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Exhibition"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Exhibition"
          ],
          "summary": "Deletes an exhibition.",
          "description": "Sample request:\r\n\r\nDELETE /api/exhibitions/1",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the exhibition to delete",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Exhibition": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "place": {
              "type": "string",
              "nullable": true
            },
            "date": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
...
        "ProblemDetails": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "nullable": true
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "status": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "detail": {
              "type": "string",
              "nullable": true
            },
            "instance": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": {
  
          }
        },
...
      }
    }
  }